Fix build warnings

+++ Noise/Expressions.cpp
@@ -51,9 +51,9 @@ float Extension::GetLoopingNoise1D(float x, float xoffset, float xsize) {
 float Extension::GetRequestNoise3D(const TCHAR* name, float x, float y, float z) {
 	NoiseRequest* r = Requests[name];
 
-	int xt = ((r->xsize + 0.f) / r->x) * x;
-	int yt = ((r->ysize + 0.f) / r->y) * y;
-	int zt = ((r->zsize + 0.f) / r->z) * z;
+	int xt = (int)(((r->xsize + 0.f) / r->x) * x);
+	int yt = (int)(((r->ysize + 0.f) / r->y) * y);
+	int zt = (int)(((r->zsize + 0.f) / r->z) * z);
 
 	//DarkEdif::MsgBox::Info(_T("inf"), _T("%f %f %f, %i %i %i"), x, y, z, xt, yt, zt);
 
@@ -63,8 +63,8 @@ float Extension::GetRequestNoise3D(const TCHAR* name, float x, float y, float z)
 float Extension::GetRequestNoise2D(const TCHAR* name, float x, float y) {
 	NoiseRequest* r = Requests[name];
 
-	int xt = ((r->xsize + 0.f) / r->x) * x;
-	int yt = ((r->ysize + 0.f) / r->y) * y;
+	int xt = (int)(((r->xsize + 0.f) / r->x) * x);
+	int yt = (int)(((r->ysize + 0.f) / r->y) * y);
 
 	//DarkEdif::MsgBox::Info(_T("inf"), _T("%f %f, %i %i"), x, y, xt, yt);
 
@@ -74,7 +74,7 @@ float Extension::GetRequestNoise2D(const TCHAR* name, float x, float y) {
 float Extension::GetRequestNoise1D(const TCHAR* name, float x) {
 	NoiseRequest* r = Requests[name];
 
-	int xt = ((r->xsize + 0.f) / r->x) * x;
+	int xt = (int)(((r->xsize + 0.f) / r->x) * x);
 
 	//DarkEdif::MsgBox::Info(_T("inf"), _T("%f, %i"), x, xt);
